Fixed #446 by blocking  popups completely in private mode.
Fix for #446, when in private mode the WKProcessPoll of the configuration is changed, this leads to a crash for popups ie window.open where the configuration is craeted by parent wkwebview. To handle this for now we are stopping popup when in private mode. The fix is done in guard statement below.