Fix #3365: Modernize NTP campaign invalidation logic.
Now we set the campaign as invalid if there is no wallpapers present
in the JSON we get from the server.

Previously we simply checked for an empty json or if the file was
smaller than 5 bytes.
This has caused regression when `schemaVersion` started to be added
to that JSON.

Current implementation should closer match what desktop does.