No Bug: Improve incremental build times
Every time a build the CurrentBundleGen plugin would write a file to the output directory, thus touching it and requiring anything that used `.current` in that module to be recompiled.

Given this plugin's lifetime is short (to be removed when we move to Xcode 14), we can optimize this by only allowing it to write once. Since the contents of this file shouldn't change we should be OK to only check for file existence and not contents