No Bug: Fix incorrect url bar focusing behaviour when opening blank tabs
As of #6150 merging, the `openBlankNewTab` was fixed to actually focus the URL bar when passing in `true` for the `attemptLocationFieldFocus` argument. However it turns out that we call this method with `true` all throughout the app even though we don't want the URL bar to focus in most of these scenarios.