Fix #6310: Do not register device with attestation server without wallet
`walletInitalized` observer is called once on initialize and again when the wallet is created. We attempt to register with the attestation server on both instead of only when the wallet is created.  We need to attempt registration on both because we need to get existing users registered after a regression so we will instead just ensure a payment ID exists before we attempt registration.

This change also fixes a potential race condition when calling `rewardsInternalInfo` which is now queued. At the moment this isn't a problem but if we eventually use mojom for Ledger this will break.