No Bug: Replace macOS app with a script for making leo asset catalogs
Running a macOS tool while archiving a iOS build is unfortunately broken in Xcode 14. It attempts to run the macOS tool from the wrong directory and fails to find it. One way around this is to use a precompiled version of this tool instead, but it will mean building and uploading them somewhere so for now the tool will be replaced by a bash script.