Fix #8661: Always allow WebKit to load popup URLs before switching tabs
There was a race condition on older devices that where WebKit would not load the popup URL before Brave attempted to select the tab. Selecting the tab would then attempt to restore it using the tab's current URL which would be `about:blank` since the web view had not made any requests yet.